Changed mutations
Decreased mutations for normal TSM (first approach) over generations. Fixed mutation value for the second approach. Since in this approach we only save the mutation with the best fitness, it is not necessary to decrease the mutation value. The idea of decreasing the value of the mutation serves to prevent it from when it reaches the best case, the continuation of the mutations with high probability make the best result regress. This never happens since the mutation is only saved if the fitness is better.